fix: format lint and move example to story
