+++ packages/react/src/components/Button/Button.stories.js
@@ -11,11 +11,6 @@ import { Add, Notification, Filter } from '@carbon/icons-react';
 import { default as Button, ButtonSkeleton } from '../Button';
 import mdx from './Button.mdx';
 import './button-story.scss';
-import {
-  ButtonRadiusExamples,
-  radiusTokenMapping,
-  radiusTokenOptions,
-} from './ButtonRadiusExamples';
 
 // Note: we explicitly define the defaultValue here, as the Button component takes `props` and forwards them
 // to the underlying `button` or `a` element, as a result storybook cannot infer the default values from the component.
@@ -29,6 +24,185 @@ const getIconFromString = (iconName) => {
   };
   return icons[iconName];
 };
+
+const radiusTokenScale = [
+  ['00', '0'],
+  ['02', '0.125rem'],
+  ['04', '0.25rem'],
+  ['08', '0.5rem'],
+  ['16', '1rem'],
+  ['24', '1.5rem'],
+  ['max', '999999px'],
+];
+
+const radiusTokenOptions = ['unset', ...radiusTokenScale.map(([name]) => name)];
+
+const radiusTokenMapping = {
+  unset: undefined,
+  ...Object.fromEntries(radiusTokenScale),
+};
+
+function toRadiusStyle({ radius, radiusSs, radiusSe, radiusEs, radiusEe }) {
+  const style = {};
+  if (radius != null) {
+    style['--cds-button-radius'] = radius;
+  }
+  if (radiusSs != null) {
+    style['--cds-button-radius-ss'] = radiusSs;
+  }
+  if (radiusSe != null) {
+    style['--cds-button-radius-se'] = radiusSe;
+  }
+  if (radiusEs != null) {
+    style['--cds-button-radius-es'] = radiusEs;
+  }
+  if (radiusEe != null) {
+    style['--cds-button-radius-ee'] = radiusEe;
+  }
+  return style;
+}
+
+const radiusStack = {
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '1.5rem',
+};
+
+const radiusGroup = {
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '0.5rem',
+};
+
+const radiusRow = {
+  display: 'flex',
+  flexWrap: 'wrap',
+  gap: '1rem',
+  alignItems: 'center',
+};
+
+const radiusLabel = {
+  color: 'var(--cds-text-secondary)',
+  fontSize: '0.875rem',
+};
+
+function ButtonRadiusExamples({
+  radius,
+  radiusSs,
+  radiusSe,
+  radiusEs,
+  radiusEe,
+  renderIcon,
+  iconDescription,
+  hasIconOnly: _hasIconOnly,
+  ...buttonProps
+}) {
+  const playgroundStyle = toRadiusStyle({
+    radius,
+    radiusSs,
+    radiusSe,
+    radiusEs,
+    radiusEe,
+  });
+  const icon = renderIcon ?? Add;
+  const iconLabel = iconDescription || 'Add';
+
+  return (
+    <div style={radiusStack}>
+      <div style={radiusGroup}>
+        <span style={radiusLabel}>Custom</span>
+        <div style={{ ...radiusRow, ...playgroundStyle }}>
+          <Button {...buttonProps} renderIcon={renderIcon}>
+            Button
+          </Button>
+          <Button
+            {...buttonProps}
+            hasIconOnly
+            renderIcon={icon}
+            iconDescription={iconLabel}
+          />
+        </div>
+      </div>
+      <div style={radiusGroup}>
+        <span style={radiusLabel}>All corners · text</span>
+        <div style={radiusRow}>
+          {radiusTokenScale.map(([name, value]) => (
+            <div key={`text-${name}`} style={toRadiusStyle({ radius: value })}>
+              <Button {...buttonProps} renderIcon={renderIcon}>
+                {name}
+              </Button>
+            </div>
+          ))}
+        </div>
+      </div>
+      <div style={radiusGroup}>
+        <span style={radiusLabel}>All corners · icon-only</span>
+        <div style={radiusRow}>
+          {radiusTokenScale.map(([name, value]) => (
+            <div key={`icon-${name}`} style={toRadiusStyle({ radius: value })}>
+              <Button
+                {...buttonProps}
+                hasIconOnly
+                renderIcon={icon}
+                iconDescription={name}
+              />
+            </div>
+          ))}
+        </div>
+      </div>
+      <div style={radiusGroup}>
+        <span style={radiusLabel}>Joined split (Combo Button)</span>
+        <div style={{ display: 'inline-flex', columnGap: 1 }}>
+          <div
+            style={toRadiusStyle({
+              radius: '999999px',
+              radiusSe: '0',
+              radiusEe: '0',
+            })}>
+            <Button {...buttonProps}>Primary action</Button>
+          </div>
+          <div
+            style={toRadiusStyle({
+              radius: '999999px',
+              radiusSs: '0',
+              radiusEs: '0',
+            })}>
+            <Button
+              {...buttonProps}
+              hasIconOnly
+              renderIcon={icon}
+              iconDescription="More"
+            />
+          </div>
+        </div>
+      </div>
+      <div style={radiusGroup}>
+        <span style={radiusLabel}>Last action end-end (AI Label footer)</span>
+        <div
+          style={{
+            ...toRadiusStyle({ radius: '0' }),
+            display: 'inline-flex',
+          }}>
+          <Button {...buttonProps} kind="ghost">
+            Cancel
+          </Button>
+          <div style={toRadiusStyle({ radiusEe: 'calc(0.5rem - 1px)' })}>
+            <Button {...buttonProps}>Confirm</Button>
+          </div>
+        </div>
+      </div>
+      <div style={radiusGroup}>
+        <span style={radiusLabel}>Nested override</span>
+        <div style={{ ...toRadiusStyle({ radius: '0' }), ...radiusRow }}>
+          <Button {...buttonProps}>Page</Button>
+          <div style={toRadiusStyle({ radius: '999999px' })}>
+            <Button {...buttonProps}>Hero</Button>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
 
 const sharedArgTypes = {
   disabled: {
