+++ packages/web-components/src/components/select/select.scss
@@ -11,7 +11,7 @@ $css--plex: true !default;
 @use '@carbon/styles/scss/spacing' as *;
 @use '@carbon/styles/scss/theme' as *;
 @use '@carbon/styles/scss/border-radius' as *;
-@use '@carbon/styles/scss/feature-flags' as *;
+@use '@carbon/styles/scss/feature-flags' as flags;
 @use '@carbon/styles/scss/components/select';
 @use '@carbon/styles/scss/colors';
 @use '@carbon/styles/scss/utilities/convert' as *;
@@ -99,7 +99,7 @@ $divider-width: 1px;
 
     // Keep pagination selects flush with the bar. Standalone v12 Select
     // rounded field chrome should not cascade into Pagination.
-    @if enabled('enable-v12-release') {
+    @if flags.enabled('enable-v12-release') {
       border: none;
       border-radius: $border-radius-00;
       background: transparent;
